Refactoring to use info popup component
